shell sort using knuth sequence

+++ sort.h
@@ -23,5 +23,5 @@ void bubble_sort(int *array, size_t size);
 int partition(int *array, ssize_t low, ssize_t high, size_t size);
 void quick_sort(int *array, size_t size);
 void quicksort(int *array, ssize_t low, ssize_t high, size_t size);
-void shell_sort(int *array, size_t size);
+void shell_sort(int *array, int size);
 #endif
